fix: Resolve additional build errors and type conflicts
- ServerAdminViewModel: Rename OnlineUserDto to AdminOnlineUserDto to avoid namespace conflict
- ServerAdminViewModel: Use UserSearchResultDto.UserId instead of .Id
- ModerationDto: Make UserMuteDto.ExpiresAt nullable to match model
- ChannelSidebar, GroupCallView, VoiceCallDashboard: Change OnLocalAudioLevel parameter from float to double

+++ src/VeaMarketplace.Client/Views/ChannelSidebar.xaml.cs
@@ -149,7 +149,7 @@ public partial class ChannelSidebar : UserControl
         }
     }
 
-    private void OnLocalAudioLevel(float level)
+    private void OnLocalAudioLevel(double level)
     {
         Dispatcher.Invoke(() =>
         {
